Renaming remote to run
